OP#9534: De-bloat BoT-SORT test suite
Remove coverage that duplicated other tests without adding signal:

- Fold six ReID model-rejection variants into one parametrize and drop
  the _build_args repeat of the same rule one call deeper.
- Drop the mocked t_miss_max finish loop and the private-attribute reset
  test; TestTrackRegistry covers the lifecycle semantics directly and
  the integration reset test covers the observable behaviour.
- Move the remaining extract_frame_rate cases into
  test_botsort_results_lite.py so one file owns that function.
- Delete the 400-seed registry equivalence suite: its oracle was
  transcribed from this branch's own pre-refactor implementation, so its
  value expired once the refactor landed; TestTrackRegistry pins the
  semantics.
