+++ tests/track/test_botsort_results_lite.py
@@ -94,6 +94,19 @@ def test_extract_frame_rate_falls_back_to_recorded() -> None:
     assert extract_frame_rate_from_metadata(metadata) == pytest.approx(25.0)
 
 
+def test_extract_frame_rate_returns_none_for_missing_video_section() -> None:
+    """Missing video section yields no FPS."""
+    assert extract_frame_rate_from_metadata({}) is None
+
+
+def test_extract_frame_rate_returns_none_for_zero_fps() -> None:
+    """Zero FPS values are treated as missing."""
+    metadata = {
+        dataformat.VIDEO: {dataformat.ACTUAL_FPS: 0, dataformat.RECORDED_FPS: 0}
+    }
+    assert extract_frame_rate_from_metadata(metadata) is None
+
+
 def test_extract_frame_rate_handles_string_encoded_fps() -> None:
     """Real .otdet files may store FPS as strings (e.g. '20.0')."""
     metadata = {
